# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice para gestão de cobranças (inicialmente integração Asaas; no futuro bancos diversos).",
        "title": "Charges Service",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
...
        "/v1/asaas/charges": {
            "get": {
                "description": "Lista cobranças (payments) do Asaas com filtros. Se company_id for informado e customer não, o serviço resolve o customer_id do Asaas via mapeamento (RPC em public) e aplica o filtro customer automaticamente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "Listar cobranças no Asaas (paginado)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da empresa (UUID) - resolve customer automaticamente",
                        "name": "company_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Elemento inicial da lista",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número de elementos da lista (max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pelo Identificador único do cliente no Asaas",
                        "name": "customer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pelo nome do grupo de cliente",
                        "name": "customerGroupName",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "UNDEFINED",
                            "BOLETO",
                            "CREDIT_CARD",
                            "PIX"
                        ],
                        "type": "string",
                        "description": "Filtrar por forma de pagamento",
                        "name": "billingType",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PENDING",
                            "RECEIVED",
                            "CONFIRMED",
                            "OVERDUE",
                            "REFUNDED",
                            "RECEIVED_IN_CASH",
                            "REFUND_REQUESTED",
                            "REFUND_IN_PROGRESS",
                            "CHARGEBACK_REQUESTED",
                            "CHARGEBACK_DISPUTE",
                            "AWAITING_CHARGEBACK_REVERSAL",
                            "DUNNING_REQUESTED",
                            "DUNNING_RECEIVED",
                            "AWAITING_RISK_ANALYSIS"
                        ],
                        "type": "string",
                        "description": "Filtrar por status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pelo Identificador único da assinatura",
                        "name": "subscription",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pelo Identificador único do parcelamento",
                        "name": "installment",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pelo identificador do seu sistema",
                        "name": "externalReference",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pela data de pagamento",
                        "name": "paymentDate",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "SCHEDULED",
                            "AUTHORIZED",
                            "PROCESSING_CANCELLATION",
                            "CANCELED",
                            "CANCELLATION_DENIED",
                            "ERROR"
                        ],
                        "type": "string",
                        "description": "Filtro para cobranças que possuam ou não nota fiscal",
                        "name": "invoiceStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pela data estimada de crédito",
                        "name": "estimatedCreditDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar recebimentos originados de um QrCode estático utilizando o id",
                        "name": "pixQrCodeId",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar registros antecipados ou não",
                        "name": "anticipated",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar registros antecipáveis ou não",
                        "name": "anticipable",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar a partir da data de criação inicial",
                        "name": "dateCreated[ge]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar até a data de criação final",
                        "name": "dateCreated[le]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar a partir da data de recebimento inicial",
                        "name": "paymentDate[ge]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar até a data de recebimento final",
                        "name": "paymentDate[le]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar a partir da data estimada de crédito inicial",
                        "name": "estimatedCreditDate[ge]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar até a data estimada de crédito final",
                        "name": "estimatedCreditDate[le]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar a partir da data de vencimento inicial",
                        "name": "dueDate[ge]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar até a data de vencimento final",
                        "name": "dueDate[le]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pelo endereço de e-mail do usuário que criou a cobrança",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar pelo identificador único da checkout",
                        "name": "checkoutSession",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AsaasPaymentsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma cobrança (payment) no Asaas para uma empresa (company_id). O serviço resolve o customer_id no Asaas via mapeamento em company.asaas_integration (RPC em public) e usa a integração ativa (is_active=true) do escritório (accounting_office_id) em iam.billing_integrations.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "Criar cobrança no Asaas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da empresa (company_id, UUID)",
                        "name": "company_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do contrato (UUID)",
                        "name": "contract_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Payload da cobrança",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AsaasCreateChargeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AsaasPaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AsaasBillingType": {
            "type": "string",
            "enum": [
                "BOLETO",
                "CREDIT_CARD",
                "PIX",
                "UNDEFINED"
            ],
            "x-enum-varnames": [
                "AsaasBillingTypeBoleto",
                "AsaasBillingTypeCreditCard",
                "AsaasBillingTypePix",
                "AsaasBillingTypeUndefined"
            ]
        },
        "model.AsaasChargeDiscount": {
            "type": "object",
            "properties": {
                "dueDateLimitDays": {
                    "type": "integer"
                },
                "type": {
                    "description": "FIXED | PERCENTAGE",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AsaasDiscountType"
                        }
                    ]
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "model.AsaasChargeFine": {
            "type": "object",
            "properties": {
                "type": {
                    "description": "FIXED | PERCENTAGE",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AsaasFineType"
                        }
                    ]
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "model.AsaasChargeInterest": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "number"
                }
            }
        },
        "model.AsaasCreateChargeRequest": {
            "type": "object",
            "properties": {
                "billingType": {
                    "description": "BOLETO | CREDIT_CARD | PIX | UNDEFINED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AsaasBillingType"
                        }
                    ]
                },
                "daysAfterDueDateToRegistrationCancellation": {
                    "description": "boleto only",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "discount": {
                    "$ref": "#/definitions/model.AsaasChargeDiscount"
                },
                "dueDate": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "externalReference": {
                    "type": "string"
                },
                "fine": {
                    "$ref": "#/definitions/model.AsaasChargeFine"
                },
                "installmentCount": {
                    "type": "integer"
                },
                "installmentValue": {
                    "type": "number"
                },
                "interest": {
                    "$ref": "#/definitions/model.AsaasChargeInterest"
                },
                "postalService": {
                    "type": "boolean"
                },
                "totalValue": {
                    "type": "number"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "model.AsaasDiscountType": {
            "type": "string",
            "enum": [
                "FIXED",
                "PERCENTAGE"
            ],
            "x-enum-varnames": [
                "AsaasDiscountTypeFixed",
                "AsaasDiscountTypePercentage"
            ]
        },
        "model.AsaasFineType": {
            "type": "string",
            "enum": [
                "FIXED",
                "PERCENTAGE"
            ],
            "x-enum-varnames": [
                "AsaasFineTypeFixed",
                "AsaasFineTypePercentage"
            ]
        },
        "model.AsaasPaymentResponse": {
            "type": "object",
            "properties": {
                "anticipable": {
                    "type": "boolean"
                },
                "anticipated": {
                    "type": "boolean"
                },
                "bankSlipUrl": {
                    "type": "string"
                },
                "billingType": {
                    "type": "string"
                },
                "canBePaidAfterDueDate": {
                    "type": "boolean"
                },
                "checkoutSession": {},
                "clientPaymentDate": {
                    "type": "string"
                },
                "confirmedDate": {
                    "type": "string"
                },
                "creditDate": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "dateCreated": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "discount": {},
                "dueDate": {
                    "type": "string"
                },
                "escrow": {},
                "estimatedCreditDate": {
                    "type": "string"
                },
                "externalReference": {
                    "type": "string"
                },
                "fine": {},
                "id": {
                    "type": "string"
                },
                "installment": {
                    "type": "string"
                },
                "installmentNumber": {
                    "type": "integer"
                },
                "interest": {},
                "interestValue": {},
                "invoiceNumber": {
                    "type": "string"
                },
                "invoiceUrl": {
                    "type": "string"
                },
                "lastBankSlipViewedDate": {},
                "lastInvoiceViewedDate": {},
                "netValue": {
                    "type": "number"
                },
                "nossoNumero": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "originalDueDate": {
                    "type": "string"
                },
                "originalValue": {},
                "paymentDate": {
                    "type": "string"
                },
                "paymentLink": {
                    "description": "duplicated field name workaround",
                    "type": "string"
                },
                "pixTransaction": {},
                "postalService": {
                    "type": "boolean"
                },
                "refunds": {},
                "status": {
                    "type": "string"
                },
                "transactionReceiptUrl": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "model.AsaasPaymentsListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AsaasPaymentResponse"
                    }
                },
                "hasMore": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "object": {
                    "description": "\"list\"",
                    "type": "string"
                },
                "offset": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
    }
}